% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiles.R
\name{checkBox}
\alias{checkBox}
\title{Check Box}
\usage{
checkBox(x, y, vert, area)
}
\arguments{
\item{x}{current X}

\item{y}{current Y}

\item{vert}{vertical box?}

\item{area}{area dataframe}
}
\value{
Bool
}
\description{
Check Box
}
\examples{
#These need updating!!
#checkBox(1,1,1,1,1,2,2)
#checkBox(0.5,0.5,1,1,1,2,2)
}
