% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diamond.r
\name{diamondLine}
\alias{diamondLine}
\title{Diamond Single Line}
\usage{
diamondLine(num_tiles, x, y)
}
\arguments{
\item{num_tiles}{number of tiles in a row}

\item{x}{start X}

\item{y}{start Y}
}
\description{
Diamond Single Line
}
\examples{
plot.new(); diamondLine(8,0,0)
}
