% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subway.R
\name{draw_tile_line}
\alias{draw_tile_line}
\title{Draw a line of subway tiles}
\usage{
draw_tile_line(num_tiles, x, y)
}
\arguments{
\item{num_tiles}{The number of tiles desired in the line}

\item{x}{The x coordinates of the starting position}

\item{y}{The y coordinates of the starting position}
}
\value{
Returns a line of line of subway tiles
}
\description{
Draw a line of subway tiles
}
\examples{
plot.new(); draw_tile_line(5, 1, 1)
}
