% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square.r
\name{draw_tile_line_square}
\alias{draw_tile_line_square}
\title{Draw a line of tiles}
\usage{
draw_tile_line_square(num_tiles, x, y)
}
\arguments{
\item{num_tiles}{The number of tiles desired in the line}

\item{x}{The x coordinates of the starting position}

\item{y}{The y coordinates of the starting position}
}
\value{
Returns a line of tiles
}
\description{
Draw a line of tiles
}
\examples{
plot.new(); draw_tile_line_square(5, 1, 1)
}
