% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiles.R
\name{drawCheckBox}
\alias{drawCheckBox}
\title{Check then draw Box}
\usage{
drawCheckBox(currentX, currentY, vert, area)
}
\arguments{
\item{currentX}{current X}

\item{currentY}{current Y}

\item{vert}{vertical?}

\item{area}{area dataframe}
}
\description{
Check then draw Box
}
\examples{
#These also need updating!
#drawCheckBox(1,1,1,1,1,2,2)
#drawCheckBox(0.5,0.5,1,1,1,2,2)
}
