% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herringbone.R
\name{herring90Fit}
\alias{herring90Fit}
\title{Herringbone 90 Degrees Fit to area}
\usage{
herring90Fit(pattern = 1, area = data.frame(x = c(1, 6), y = c(1, 6)))
}
\arguments{
\item{pattern}{which pattern starts the tiling}

\item{area}{dataframe area}
}
\description{
Herringbone 90 Degrees Fit to area
}
\examples{
plot.new(); herring90Fit()
plot.new(); herring90Fit(2)
}
