% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subway.R
\name{stack_tile_line}
\alias{stack_tile_line}
\title{Stack tile lines on top of each other}
\usage{
stack_tile_line(num_tiles, num_verts, x, y, offset)
}
\arguments{
\item{num_tiles}{The number of tiles you would like in the line}

\item{num_verts}{The number of vertical rows you would like}

\item{x}{The x coordinates of the starting line}

\item{y}{The y coordinates of the starting line}

\item{offset}{the offset of the second tile line}
}
\value{
Returns a tiling pattern
}
\description{
Stack tile lines on top of each other
}
\examples{
plot.new(); stack_tile_line(4, 6, 1, 1, 0.5)
}
