% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diamond.r
\name{stackTiles}
\alias{stackTiles}
\title{Diamond Multi Line}
\usage{
stackTiles(num_tiles, vert, x, y)
}
\arguments{
\item{num_tiles}{number of tiles in a row}

\item{vert}{number of tile rows stacked}

\item{x}{start X}

\item{y}{start Y}
}
\description{
Diamond Multi Line
}
\examples{
stackTiles(8,8,0,0)
}
