% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiles.R
\name{tileShift}
\alias{tileShift}
\title{Tile Placement function}
\usage{
tileShift(shape, x, y)
}
\arguments{
\item{shape}{shape specifications}

\item{x}{x co-ordinate}

\item{y}{y co-ordinate}
}
\description{
Tile Placement function
}
